feat!: 점수 등급(A~F) 제거, 0~100 정수 점수로 환원
사용자 지정(2026-08-15). 화면이 `87점 / 100`을 그리는데 등급 문자가 따로
필요 없었고, 소수점은 자릿수가 흔들려 레이아웃이 밀렸다. 발표 점수에서
0.4점 차이는 사용자에게 의미가 없다.

- utils/score_grade.py 삭제. 응답의 grades 필드 제거.
- overall_scores와 words_detail[].accuracy_score 모두 0~100 정수.
  한 화면에 "종합 58"과 "구축의 52.4"가 같이 뜨면 기준을 알 수 없다.
- DB 컬럼 Float -> Integer. 다만 SQLite는 선언 타입을 강제하지 않아
  8/15 이전 행에는 소수가 그대로 남는다. 조회할 때 한 번 더 정수로
  거르므로(_int_score) 옛 기록만 소수로 보이지는 않는다.
- 점수 없음(None)은 0으로 채우지 않는다. 평가 실패와 0점은 다르고,
  0으로 채우면 F를 받은 것처럼 보인다.

문서: 프론트 가이드와 API 명세서에 "화면에는 등급만 표기, 숫자 점수는
쓰지 마세요"가 적혀 있어 그대로 두면 프론트가 없어진 필드를 읽는다.
등급 규칙을 걷어내고 정수 표기로 바꿨다. 하이라이팅·감점 요인 사용법도
같은 문서에 추가했고, 스프링 연동 요청서에 전/후 대조표를 넣었다.

pytest 428건 통과.

+++ speako-ai-server/src/db/models.py
@@ -1,6 +1,6 @@
 from datetime import datetime, timezone
 
-from sqlalchemy import Column, Integer, String, Text, JSON, DateTime, Float, ForeignKey
+from sqlalchemy import Column, Integer, String, Text, JSON, DateTime, ForeignKey
 from sqlalchemy.orm import relationship
 
 from db.database import Base
@@ -85,10 +85,14 @@ class PronunciationEvaluation(Base):
     # 슬라이드별로 나눠 녹음한 경우 그 슬라이드 번호. 대본 전체를 한 번에 녹음했으면 None.
     # (코칭 내역에서 "3번 슬라이드 87점"처럼 구분하려면 이게 남아 있어야 한다)
     slide_number = Column(Integer, nullable=True)
-    accuracy_score = Column(Float, nullable=True)
-    fluency_score = Column(Float, nullable=True)
-    completeness_score = Column(Float, nullable=True)
-    pronunciation_score = Column(Float, nullable=True)
+    # 0~100 정수. Azure는 소수를 주지만 저장 전에 반올림한다(main._round_scores_in_place).
+    # ⚠️ 2026-08-15 이전에 저장된 행에는 소수(87.4)가 그대로 남아 있다. SQLite는 선언한
+    #    타입을 강제하지 않아 그 값이 그대로 읽히므로, 조회할 때 한 번 더 정수로 거른다
+    #    (main._int_score). 옛 기록만 소수로 보이면 안 된다.
+    accuracy_score = Column(Integer, nullable=True)
+    fluency_score = Column(Integer, nullable=True)
+    completeness_score = Column(Integer, nullable=True)
+    pronunciation_score = Column(Integer, nullable=True)
     words_detail = Column(JSON, nullable=True)
     # 평가 기준으로 쓴 원본 대본과, Azure가 실제로 인식한 텍스트. 결과 화면에서 둘을 나란히 놓고
     # 어디를 다르게 읽었는지 비교한다(피그마 Feedback Page: 원본 텍스트 ↔ 인식 텍스트).
